requirements.txt
aerich==0.7.1
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dictdiffer==0.9.0
ecdsa==0.18.0
fastapi==0.85.1
frozenlist==1.3.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
multidict==6.0.2
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pypika-tortoise==0.1.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.6
tortoise-orm==0.19.2
typing-extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.3
yarl==1.8.1
